+++ website/src/app/[lang]/[country]/(website)/transparency/[currency]/page.tsx
@@ -1,10 +1,14 @@
 import { DefaultPageProps } from '@/app/[lang]/[country]';
-import { getStats } from '@/app/[lang]/[country]/(website)/transparency/[currency]/get-stats';
 import TransparencyCharts from '@/app/[lang]/[country]/(website)/transparency/[currency]/transparency-charts';
 import { ValidCountry, WebsiteLanguage } from '@/i18n';
 import { BaseContainer, Stats } from '@socialincome/ui';
+import {firestoreAdmin} from "@/firebase/admin";
+import { ContributionStatsCalculator } from '@socialincome/shared/src/utils/stats/ContributionStatsCalculator';
+import { PaymentStatsCalculator } from '@socialincome/shared/src/utils/stats/PaymentStatsCalculator';
+import Test from "@/app/[lang]/[country]/(website)/transparency/[currency]/test";
 
 export const generateStaticParams = () => ['USD', 'CHF'].map((currency) => ({ currency: currency.toLowerCase() }));
+export const revalidate = 60 // once an hour
 
 export type TransparencyPageProps = {
 	params: {
@@ -15,6 +19,16 @@ export type TransparencyPageProps = {
 } & DefaultPageProps;
 
 export default async function Page(props: TransparencyPageProps) {
+	const getStats = async (currency: string) => {
+		console.log('retrieving stats');
+		const contributionCalculator = await ContributionStatsCalculator.build(firestoreAdmin, currency);
+		const contributionStats = contributionCalculator.allStats();
+		const paymentCalculator = await PaymentStatsCalculator.build(firestoreAdmin, currency);
+		const paymentStats = paymentCalculator.allStats();
+		console.log('retrieved stats');
+		return { contributionStats, paymentStats };
+	};
+
 	const { contributionStats, paymentStats } = await getStats(props.params.currency);
 
 	return (
@@ -26,7 +40,7 @@ export default async function Page(props: TransparencyPageProps) {
 					<Stats.Stat.Item variant="value">{contributionStats.totalContributions}</Stats.Stat.Item>
 				</Stats.Stat>
 			</Stats>
-			<TransparencyCharts contributionStats={contributionStats} paymentStats={paymentStats} {...props} />
+			<TransparencyCharts contributionStats={contributionStats} paymentStats={paymentStats} />
 		</BaseContainer>
 	);
 }
