+++ website/src/app/[lang]/[country]/(website)/transparency/[currency]/test.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { ContributionStats } from '@socialincome/shared/src/utils/stats/ContributionStatsCalculator';
+
+interface TestParam {
+    contributionStats: ContributionStats
+}
+
+export default async function Test(testParam: TestParam) {
+    return <div>{testParam.contributionStats.totalContributions}</div>
+}
